TST: refactors environment.set_date test TST: refactors tests/unit/test_environment TST: refactors environment_fixtures DOC: fix doc typo in decimal_degrees_to_arc_seconds TST: fixes example_spaceport_env legacy references TST: fixes example_plain_env legacy references TST: Rollback teardown logic in unit env test

+++ rocketpy/environment/environment.py
@@ -3696,7 +3696,7 @@ class Environment:
         -------
         degrees : float
             The degrees.
-        arc_minutes : float
+        arc_minutes : int
             The arc minutes. 1 arc-minute = (1/60)*degree
         arc_seconds : float
             The arc Seconds. 1 arc-second = (1/3600)*degree
